requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2     
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.3        
future==0.18.2
idna==3.4
multidict==6.0.2
pefile==2022.5.30        
playsound==1.2.2
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
python-dateutil==2.8.2
python-dotenv==0.21.0
pywin32-ctypes==0.2.0
six==1.16.0
twitchAPI==3.0.1
typing_extensions==4.4.0
yarl==1.8.1